feat(alpha-groups): add Alpha Groups tab to admin dashboard
- Introduced a new Alpha Groups tab for monitoring and managing alpha group dynamics.
- Integrated statistics display including total members, invites, and tier distribution.
- Implemented fetching of alpha group statistics and configuration from the API.
- Added loading and error handling states for improved user experience.
- Removed the deprecated alpha-groups page component.
